NSOF-8035 ci: add our custom build workflow script
Build the otelcol-contrib debian package and upload it as an artifact.

Goreleaser defaults to validate that releases are only done from tagged
versions therefore run it with --skip-validate so that we can build from
the branch; however, since this is creating the same release every time,
the release has to be manually be deleted before triggering a build.

So for "official" builds we should tag before building.